Simplify text coordinate calculations
- tile-labeler (dependency): return font scalar with labelPos,
  normalize SDF rects (texture coordinates), add w, h to charPos
  and scale by font scalar
- main.js: loadAtlas now returns the texture sampler only. (width,
  height properties are no longer needed since SDF texture coords
  are pre-normalized)
- text/vert.glsl: compute SDF taper width from font scalar and
  screen pixel ratio; update to new tile-labeler attribute buffer
  structure
- text/frag.glsl: use taper width from vertex shader, and drop
  normalization of SDF texture coordinates
- text/program.js: drop setAtlas function (use uniform setter
  directly) and incorporate loader.js code
