+++ package.json
@@ -30,7 +30,7 @@
   },
   "homepage": "https://github.com/GlobeletJS/tile-gl#readme",
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^19.0.0",
+    "@rollup/plugin-commonjs": "^19.0.2",
     "@rollup/plugin-node-resolve": "^13.0.0",
     "eslint": "^7.30.0",
     "eslint-config-globeletjs": "^0.0.6",
@@ -38,7 +38,7 @@
   },
   "dependencies": {
     "earcut": "^2.2.3",
-    "tile-labeler": "^0.2.3"
+    "tile-labeler": "^0.3.0"
   },
   "sideEffects": false
 }
