+++ src/text/frag.glsl
@@ -1,18 +1,15 @@
 precision highp float;
 
 uniform sampler2D sdf;
-uniform vec2 sdfDim;
 
 varying vec4 fillStyle;
 varying vec2 texCoord;
+varying float taperWidth;
 
 void main() {
-  float sdfVal = texture2D(sdf, texCoord / sdfDim).a;
-  // Find taper width: ~ dScreenPixels / dTexCoord
-  float screenScale = 1.414 / length(fwidth(texCoord));
-  float screenDist = screenScale * (191.0 - 255.0 * sdfVal) / 32.0;
+  float sdfVal = texture2D(sdf, texCoord).a;
+  float screenDist = taperWidth * (191.0 - 255.0 * sdfVal) / 32.0;
 
-  // TODO: threshold 0.5 looks too pixelated. Why?
-  float alpha = smoothstep(-0.8, 0.8, -screenDist);
+  float alpha = smoothstep(-0.707, 0.707, -screenDist);
   gl_FragColor = fillStyle * alpha;
 }
