--- src/text/loader.js
@@ -1,26 +1,0 @@
-export function initTextLoader(context, constructVao) {
-  const { initQuad, initAttribute } = context;
-
-  const quadPos = initQuad({ x0: 0.0, y0: 0.0, x1: 1.0, y1: 1.0 });
-
-  const attrInfo = {
-    labelPos: { numComponents: 2 },
-    charPos: { numComponents: 3 },
-    sdfRect: { numComponents: 4 },
-    tileCoords: { numComponents: 3 },
-    color: { numComponents: 4 },
-    opacity: { numComponents: 1 },
-  };
-
-  return function(buffers) {
-    const attributes = Object.entries(attrInfo).reduce((d, [key, info]) => {
-      const data = buffers[key];
-      if (data) d[key] = initAttribute(Object.assign({ data }, info));
-      return d;
-    }, { quadPos });
-
-    const vao = constructVao({ attributes });
-
-    return { vao, instanceCount: buffers.labelPos.length / 2 };
-  };
-}
